Add files via upload
This is a C++ Member Registration System for a restaurant. 

This project is a prototype for a simple front-end and back-end database desktop application, which allows the user to create and store member data in the MS SQL database. It also provides a verification function to check whether data input of existing members can match the data in the MS SQL database.

If you are interested, please download all files and open the .sln file to launch the program.

The Microsoft SQL server can be reset to be connected to your own server on Microsoft Visual Studio.